chore(lineup): refactor some routes
